backend/requirements.txt
django==5.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
google-api-python-client==2.100.0
django-cors-headers==4.3.0
drf-spectacular==0.26.5
celery==5.3.4
redis==5.0.1
Pillow==10.0.0
pytest==7.4.2
pytest-django==4.5.2
